reduce ValueTable width to 32 bits
--------------------------------------------------
Results of dev vs base (1+0.01, 1t, 1MB, UHO_Lichess_4852_v1.epd):
Elo: 1.80 +/- 1.46, nElo: 3.36 +/- 2.72
LOS: 99.22 %, DrawRatio: 51.94 %, PairsRatio: 1.04
Games: 62574, Wins: 16432, Losses: 16108, Draws: 30034, Points: 31449.0 (50.26 %)
Ptnml(0-2): [476, 6886, 16252, 7184, 489], WL/DD Ratio: 1.04
LLR: 2.89 (100.1%) (-2.25, 2.89) [0.00, 3.00]
--------------------------------------------------

+++ lib/search/engine.rs
@@ -1360,7 +1360,7 @@ impl Engine {
     pub fn reset(&mut self) {
         let local: &[SyncUnsafeCell<LocalData>] = unsafe { &*(&raw mut *self.local as *const _) };
 
-        let vt: &[SyncUnsafeCell<MaybeUninit<u64>>] =
+        let vt: &[SyncUnsafeCell<MaybeUninit<u32>>] =
             unsafe { &*(&raw mut *self.shared.vt as *const _) };
         let tt: &[SyncUnsafeCell<MaybeUninit<u64>>] =
             unsafe { &*(&raw mut *self.shared.tt as *const _) };
@@ -1371,7 +1371,7 @@ impl Engine {
         self.executor.execute(move |idx| unsafe {
             let offset = idx * vt_chunk_size;
             let len = vt.len().saturating_sub(offset).min(vt_chunk_size);
-            let ptr = vt.as_ptr().add(offset) as *mut MaybeUninit<u64>;
+            let ptr = vt.as_ptr().add(offset) as *mut MaybeUninit<u32>;
             fill_zeroes(slice::from_raw_parts_mut(ptr, len));
 
             let offset = idx * tt_chunk_size;
